refactor: created the single and array folders in the recordConversion directory
