+++ src/custom/recordConversion/array/dataToStoredRecords.ts
@@ -1,5 +1,5 @@
-import { StoredRecord } from '../..';
-import { dataObjectToStoredRecord } from './dataObjectToStoredRecord';
+import { StoredRecord } from '../../../index';
+import { dataObjectToStoredRecord } from '../single/dataObjectToStoredRecord';
 
 export function dataToStoredRecords(dataRecords: object[]): StoredRecord[] {
     return dataRecords.map(dataRecord => dataObjectToStoredRecord(dataRecord))
