+++ src/custom/recordConversion/array/liveRecordsToData.ts
@@ -1,5 +1,5 @@
 import { Record } from 'neo4j-driver-core';
-import { liveRecordToData } from './liveRecordToData';
+import { liveRecordToData } from '../single/liveRecordToData';
 
 export function liveRecordsToData(records: Record[]): object[] {
     return records.map(record => liveRecordToData(record))
