+++ src/custom/recordConversion/array/storedRecordsToData.ts
@@ -1,5 +1,5 @@
-import {storedRecordToData} from "./storedRecordToData";
-import { StoredRecord } from '../..';
+import {storedRecordToData} from "../single/storedRecordToData";
+import { StoredRecord } from '../../../index';
 
 export function storedRecordsToData(records: StoredRecord[]): object[] {
     return records.map(record => storedRecordToData(record))
