+++ src/custom/recordConversion/single/dataObjectToLiveRecord.ts
@@ -1,6 +1,6 @@
 import traverse = require('traverse');
 import { Record } from 'neo4j-driver-core';
-import { neo4jIntDataToLive } from '../neo4jInts/neo4jIntDataToLive';
+import { neo4jIntDataToLive } from '../../neo4jInts/neo4jIntDataToLive';
 
 export function dataObjectToLiveRecord(dataObject: object): Record {
     const keys = Object.keys(dataObject)
