+++ src/custom/response/dataToLive.ts
@@ -1,4 +1,4 @@
-import { dataToLiveRecords } from '../recordConversion/dataToLiveRecords';
+import { dataToLiveRecords } from '../recordConversion/array/dataToLiveRecords';
 import { LiveResponse } from '../types/LiveResponse';
 
 export function dataToLive(data: object[]): LiveResponse {
