+++ src/custom/response/liveToData.ts
@@ -1,5 +1,5 @@
 import { LiveResponse } from '../types/LiveResponse';
-import { liveRecordsToData } from '../recordConversion/liveRecordsToData';
+import { liveRecordsToData } from '../recordConversion/array/liveRecordsToData';
 
 export function liveToData(liveResponse: LiveResponse): object[] {
     const {records} = liveResponse
