+++ src/custom/response/storedToData.ts
@@ -1,5 +1,5 @@
 import { StoredResponse } from '../types/StoredResponse';
-import { storedRecordsToData } from '../recordConversion/storedRecordsToData';
+import { storedRecordsToData } from '../recordConversion/array/storedRecordsToData';
 
 export function storedToData(storedResponse: StoredResponse): object[] {
     const {records} = storedResponse
